avoid n+1 querries (orientation_date still a problem) + add StatsConcern
